move about and textForm in pages fix bg color not cover page issue
