+++ src/index.css
@@ -1,5 +1,12 @@
-body {
+html,
+body,
+#root {
+  height: 100%;
   margin: 0;
+  padding: 0;
+}
+
+body {
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
@@ -12,6 +19,6 @@ code {
     monospace;
 }
 
-.accordion-button::after{
-  filter:invert(1);
+.accordion-button::after {
+  filter: invert(1);
 }
